Clean up extension tests, fix broken PackageDirectoryBuilderTests
