+++ test/DynamoCoreTests/ExtensionTests.cs
@@ -6,13 +6,11 @@ using System.Threading.Tasks;
 using NUnit.Framework;
 using Dynamo.Extensions;
 using System.IO;
-using Dynamo.PackageManager;
-using System.Reflection;
 using Dynamo.Models;
 using Moq;
 using Dynamo.Interfaces;
 
-namespace PackageManagerTests
+namespace Dynamo.Tests
 {
     class ExtensionTests
     {
@@ -38,16 +36,6 @@ namespace PackageManagerTests
         public void DirectoryExists()
         {
             Assert.IsTrue(Directory.Exists(extensionsPath));
-        }
-
-        [Test]
-        public void ExtensionsAreExtracted()
-        {
-            var extensionManager = new ExtensionManager();
-            var extensions = extensionManager.ExtensionLoader.LoadDirectory(extensionsPath);
-            Assert.Greater(extensions.Count(), 0);
-
-            Assert.AreEqual(extensions.OfType<PackageManagerExtension>().Count(), 1);
         }
 
         [Test]
